# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriSmart – Intelligent Farm Management and Decision Support System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms/{farmId}/crops:
    get:
      operationId: listCrops
      tags: [crops]
      summary: List crops for a farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of crops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crop"
    post:
      operationId: createCrop
      tags: [crops]
      summary: Add a crop to a farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropInput"
      responses:
        "201":
          description: Created crop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
  /crops/{id}:
    get:
      operationId: getCrop
      tags: [crops]
      summary: Get a crop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Crop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
        "404":
          description: Not found
    put:
      operationId: updateCrop
      tags: [crops]
      summary: Update a crop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CropInput"
      responses:
        "200":
          description: Updated crop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
        "404":
          description: Not found
    delete:
      operationId: deleteCrop
      tags: [crops]
      summary: Delete a crop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Expenses ────────────────────────────────────────────────────────────────
...
  /disease-detections:
    get:
      operationId: listDiseaseDetections
      tags: [disease]
      summary: List disease detection history
      parameters:
        - name: farmId
          in: query
          schema:
            type: number
        - name: cropId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Detection history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiseaseDetection"
    post:
      operationId: createDiseaseDetection
      tags: [disease]
      summary: Detect disease from a crop leaf image (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiseaseDetectionInput"
      responses:
        "201":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseDetection"
  /disease-detections/{id}:
    get:
      operationId: getDiseaseDetection
      tags: [disease]
      summary: Get a disease detection by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Detection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiseaseDetection"
        "404":
          description: Not found

  # ─── Fertilizer Recommendations ──────────────────────────────────────────────
...
  /admin/users/{id}/role:
    put:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Not found
components:
  schemas:
...
    AdminUser:
      type: object
      required: [id, clerkId, email, role, createdAt]
      properties:
        id:
          type: number
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [farmer, admin]
        createdAt:
          type: string
          format: date-time
        farmsCount:
          type: number
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [farmer, admin]

    # ── Farms ──────────────────────────────────────────────────────────────────
...
    Crop:
      type: object
      required: [id, farmId, name, status, plantingDate, createdAt]
      properties:
        id:
          type: number
        farmId:
          type: number
        name:
          type: string
        variety:
          type: ["string", "null"]
        status:
          type: string
          enum: [growing, harvested, failed, planned]
        plantingDate:
          type: string
          format: date
        harvestDate:
          type: ["string", "null"]
          format: date
        expectedYield:
          type: ["number", "null"]
          description: Expected yield in kg
        actualYield:
          type: ["number", "null"]
          description: Actual yield in kg
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CropInput:
      type: object
      required: [name, plantingDate, status]
      properties:
        name:
          type: string
          minLength: 1
        variety:
          type: string
        status:
          type: string
          enum: [growing, harvested, failed, planned]
        plantingDate:
          type: string
          format: date
        harvestDate:
          type: string
          format: date
        expectedYield:
          type: number
        actualYield:
          type: number
        notes:
          type: string

    # ── Expenses ──────────────────────────────────────────────────────────────
...
    DiseaseDetection:
      type: object
      required: [id, farmId, diseaseName, confidence, symptoms, treatment, imageUrl, createdAt]
      properties:
        id:
          type: number
        farmId:
          type: number
        cropId:
          type: ["number", "null"]
        imageUrl:
          type: string
        diseaseName:
          type: string
        confidence:
          type: number
          description: Confidence score 0–1
        symptoms:
          type: string
        treatment:
          type: string
        isHealthy:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DiseaseDetectionInput:
      type: object
      required: [farmId, imageBase64]
      properties:
        farmId:
          type: number
        cropId:
          type: number
        imageBase64:
          type: string
          description: Base64-encoded crop leaf image

    # ── Fertilizer ──────────────────────────────────────────────────────────
...